backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
beautifulsoup4==4.13.4
    # via -r requirements.in
build==1.3.0
    # via pip-tools
certifi==2025.7.14
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==3.4.2
    # via
    #   -r requirements.in
    #   requests
click==8.2.1
    # via pip-tools
greenlet==3.2.3
    # via
    #   -r requirements.in
    #   playwright
idna==3.10
    # via
    #   -r requirements.in
    #   requests
numpy==2.3.2
    # via pandas
packaging==25.0
    # via build
pandas==2.2.3
    # via -r requirements.in
pip-tools==7.5.0
    # via -r requirements.in
playwright==1.54.0
    # via -r requirements.in
pyee==13.0.0
    # via
    #   -r requirements.in
    #   playwright
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
requests==2.32.4
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
soupsieve==2.7
    # via
    #   -r requirements.in
    #   beautifulsoup4
typing-extensions==4.14.1
    # via
    #   -r requirements.in
    #   beautifulsoup4
    #   pyee
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   -r requirements.in
    #   requests
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
langchain
langchain-community
psycopg2-binary
python-dotenv
